# perf: implement cache-friendly data layout optimizations